+++ src/types/mod.rs
@@ -1,5 +1,3 @@
 mod multiplier;
-mod std_dev;
 
 pub use multiplier::*;
-pub use std_dev::*;